@@ -41,24 +41,24 @@
     "zustand": "^4.4.0"
   },
   "devDependencies": {
-    "@types/eslint": "^8.21.3",
-    "@types/node": "^18.15.5",
+    "@types/eslint": "^8.37.0",
+    "@types/node": "^18.15.11",
     "@types/prettier": "^2.7.2",
-    "@types/react": "^18.0.28",
+    "@types/react": "^18.0.33",
     "@types/react-dom": "^18.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.56.0",
-    "@typescript-eslint/parser": "^5.56.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@typescript-eslint/parser": "^5.57.1",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.36.0",
+    "eslint": "^8.37.0",
     "eslint-config-next": "^13.4.12",
     "postcss": "^8.4.21",
     "postcss-preset-mantine": "^1.7.0",
     "postcss-simple-vars": "^7.0.1",
-    "prettier": "^2.8.6",
+    "prettier": "^2.8.7",
     "prettier-plugin-tailwindcss": "^0.4.1",
     "tailwind-custom-utilities": "^1.0.3",
     "tailwindcss": "^3.3.3",
-    "typescript": "^5.0.2"
+    "typescript": "^5.0.3"
   },
   "ct3aMetadata": {
     "initVersion": "7.10.2"
