@@ -1,3 +1,4 @@
+"use client"
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import { useResponse } from "@/hooks";
 import { Badge, Paper, Text, Skeleton, Title } from "@mantine/core";
@@ -5,10 +6,10 @@ import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import { Actions } from ".";
 import Typewriter from "typewriter-effect";
-import { useRouter } from "next/router";
+import { useSearchParams } from "next/navigation";
 
 const H1 = () => (
-  <Title order={1} className="w-full" size="md" mt="xl">
+  <Title order={1} className="w-full" mt="lg" mb="lg">
     <Typewriter
       onInit={(typewriter) => {
         typewriter
@@ -29,13 +30,12 @@ const IntialView = () => {
       withBorder
       radius="lg"
       className="w-full gap-2 flex-col-start"
-      p="xl"
-      shadow="md"
+      p={{ base: "sm", md: "lg" }}
     >
       <H1 />
       <div className="mb-4 ">
-        <Text className="text-xl font-bold" c="dimmed" mb="xs">
-          What a regular dictionary can look up
+        <Text className="ml-1 text-xl font-bold" c="dimmed" mb="xs">
+          Regular dictionary
         </Text>
         <div>
           <Badge tt="inherit" p="lg" size="lg">
@@ -44,8 +44,8 @@ const IntialView = () => {
         </div>
       </div>
       <div className="mb-4 ">
-        <Text className="text-xl font-bold" c="dimmed" mb="xs">
-          What KamousAI can look up
+        <Text className="ml-1 text-xl font-bold" c="dimmed" mb="xs">
+          AI dictionary
         </Text>
         <div className="grid grid-cols-2 gap-4 md:grid-cols-3 ">
           {[
@@ -56,7 +56,7 @@ const IntialView = () => {
             "Word or word",
             "Misspelled words",
           ].map((s) => (
-            <Badge key={s} p="lg" tt="inherit" size="lg">
+            <Badge key={s} p="lg" tt="inherit" size="lg" w="100%">
               {s}
             </Badge>
           ))}
@@ -90,18 +90,19 @@ const ActionResponse = () => {
 export const Response = () => {
   const definition = useResponse((s) => s.definition);
   const status = useResponse((s) => s.status);
-  const { query } = useRouter();
+  const query = useSearchParams();
+  const term = query?.get("term");
   return (
     <div>
-      {status !== "loading" && !query.term && <IntialView />}
-      {definition && !!query.term && (
+      {status !== "loading" && !term && <IntialView />}
+      {definition && !!term && (
         <Paper
-          withBorder
           radius="lg"
-          py="lg"
-          px="xs"
-          className="prose mx-auto w-full max-w-3xl md:text-lg font-medium tracking-wide prose-thead:bg-indigo-200/60 rounded-t" 
+          py="xl"
+          px={{base: "xs", md: "lg"}}
+          className="prose mx-auto w-full max-w-2xl md:text-lg font-medium tracking-wide prose-thead:bg-white/60" 
           shadow="md"
+          bg="dark"
         >
           <ReactMarkdown remarkPlugins={[[remarkGfm, { singleTilde: false }]]}>
             {definition}
